Move project files reader
