+++ src/main/java/tech/jhipster/lite/module/domain/TemplatedFile.java
@@ -1,9 +1,9 @@
 package tech.jhipster.lite.module.domain;
 
 import java.nio.file.Path;
-import tech.jhipster.lite.common.domain.ProjectFilesReader;
 import tech.jhipster.lite.error.domain.Assert;
 import tech.jhipster.lite.module.domain.properties.JHipsterProjectFolder;
+import tech.jhipster.lite.projectfile.domain.ProjectFilesReader;
 
 public class TemplatedFile {
