+++ src/main/java/tech/jhipster/lite/npm/infrastructure/secondary/FileSystemNpmVersions.java
@@ -4,13 +4,13 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.springframework.stereotype.Repository;
 import tech.jhipster.lite.common.domain.Enums;
-import tech.jhipster.lite.common.domain.ProjectFilesReader;
 import tech.jhipster.lite.error.domain.Assert;
 import tech.jhipster.lite.npm.domain.NpmPackageName;
 import tech.jhipster.lite.npm.domain.NpmVersion;
 import tech.jhipster.lite.npm.domain.NpmVersionSource;
 import tech.jhipster.lite.npm.domain.NpmVersions;
 import tech.jhipster.lite.npm.domain.UnknownNpmPackageException;
+import tech.jhipster.lite.projectfile.domain.ProjectFilesReader;
 
 @Repository
 public class FileSystemNpmVersions implements NpmVersions {
