+++ src/main/java/tech/jhipster/lite/projectfile/domain/ProjectFilesReader.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.common.domain;
+package tech.jhipster.lite.projectfile.domain;
 
 public interface ProjectFilesReader {
   String readString(String path);
